+++ app/lib/shared/versions.dart
@@ -21,7 +21,7 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// Make sure that at least two versions are kept here as the next candidates
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
-  '2021.02.16', // The current [runtimeVersion].
+  '2021.02.23', // The current [runtimeVersion].
   '2021.02.12',
   '2021.01.29',
 ];
@@ -55,8 +55,8 @@ bool shouldGCVersion(String version) =>
 final String runtimeSdkVersion = '2.10.0';
 final String toolStableDartSdkVersion = '2.10.5';
 final String toolStableFlutterSdkVersion = '1.22.6';
-final String toolPreviewDartSdkVersion = '2.12.0-259.9.beta';
-final String toolPreviewFlutterSdkVersion = '1.26.0-17.5.pre';
+final String toolPreviewDartSdkVersion = '2.12.0-259.15.beta';
+final String toolPreviewFlutterSdkVersion = '1.26.0-17.6.pre';
 
 // Value comes from package:pana.
 final String panaVersion = pana.packageVersion;